+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Legenden
+ */
+$GLOBALS['TL_LANG']['tl_settings']['fernschach_legend']       = 'Fernschach-Verwaltung';
+
+/**
+ * Felder
+ */
+$GLOBALS['TL_LANG']['tl_settings']['fernschach_resetActive'] = array('Globalen Reset-Buchungsdatensatz aktivieren','Aktiviert einen Reset-Buchungsdatensatz bei jedem Spieler. Ist diese Option nicht aktiviert, werden evtl. vorhandene Reset-Buchungsdatensätze beim Aufruf der Buchungen eines Spielers gelöscht.');
+$GLOBALS['TL_LANG']['tl_settings']['fernschach_resetDate'] = array('Buchungsdatum','Datum und Uhrzeit des Reset-Buchungsdatensatzes.');
+$GLOBALS['TL_LANG']['tl_settings']['fernschach_resetSaldo'] = array('Saldo','Betrag für den Reset-Buchungsdatensatz.');